+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 0.38.0 (Unreleased)
+ENHANCEMENTS:
+* managed-kubernetes: allow to create cluster with KMS provider for secrets encryption.
+
 ## 0.37.0 (April 16, 2020)
 ENHANCEMENTS:
 * storage: support custom acl grants for `yandex_storage_bucket`